requirements.txt
feedparser>=6.0.0
requests>=2.28.0
python-dotenv>=0.20.0
openai>=1.0.0
